Secured views for adding, updating and deleting views to prevent site users to enter admin user pages

+++ books/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Book, Sub_Category
@@ -73,8 +74,13 @@ def book_detail(request, book_id):
     return render(request, 'books/book_detail.html', context)
 
 
+@login_required
 def add_book(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can access this page.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = BookForm(request.POST, request.FILES)
         if form.is_valid():
@@ -94,8 +100,13 @@ def add_book(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_book(request, product_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can access this page.')
+        return redirect(reverse('home'))
+
     book = get_object_or_404(Book, pk=product_id)
     if request.method == 'POST':
         form = BookForm(request.POST, request.FILES, instance=book)
@@ -118,8 +129,13 @@ def edit_book(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_book(request, product_id):
     """ Delete a book from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can access this page.')
+        return redirect(reverse('home'))
+
     book = get_object_or_404(Book, pk=product_id)
     book.delete()
     messages.success(request, 'Book deleted!')
